web/src/requirements.txt
click==6.7
dateutils==0.6.6
Flask==0.12.2
Flask-Login==0.4.0
Flask-WTF==0.14.2
fuzzywuzzy==0.15.1
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
nameparser==0.5.3
passlib==1.7.1
pymongo==3.5.1
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2017.2
redis==2.10.6
six==1.14.0
Werkzeug==0.12.2
WTForms==2.1
